+++ polizador/secretariador/forms/vehiculoform.py
@@ -1,6 +1,13 @@
 from django import forms
 from secretariador.models import Vehiculo
 from secretariador.forms.mixins import BaseFormMixin
+from secretariador.views.ajaxviews import (
+    ResolucionWidget,
+    ComisionadoWidget,
+    VehiculoWidget,
+    DecretoWidget,
+    )
+from carga.views.ajaxviews import empresawidget, aseguradorawidget
 
 class VehiculoForm(BaseFormMixin, forms.ModelForm):
     class Meta:
@@ -11,6 +18,11 @@ class VehiculoForm(BaseFormMixin, forms.ModelForm):
             "vehiculo_patente",
             "vehiculo_poliza",
             "vehiculo_poliza_aseguradora",
+            "vehiculo_titular_agente",
+            "vehiculo_titular_empresa",
+            "vehiculo_n_motor",
+            "vehiculo_n_chasis",
+            "vehiculo_modelo",
         )
         widgets = {
             "vehiculo_caracter":forms.Select(attrs={
@@ -25,7 +37,22 @@ class VehiculoForm(BaseFormMixin, forms.ModelForm):
             "vehiculo_poliza":forms.TextInput(attrs={
                 "class":"form-control",
                 }),
-            "vehiculo_poliza_aseguradora":forms.Select(attrs={
+            "vehiculo_poliza_aseguradora":aseguradorawidget(attrs={
                 "class":"form-control customSelect2",
+                }),
+            "vehiculo_titular_agente":ComisionadoWidget(attrs={
+                "class":"form-control customSelect2",
+                }),
+            "vehiculo_titular_empresa":empresawidget(attrs={
+                "class":"form-control customSelect2",
+                }),
+            "vehiculo_n_motor":forms.TextInput(attrs={
+                "class":"form-control",
+                }),
+            "vehiculo_n_chasis":forms.TextInput(attrs={
+                "class":"form-control",
+                }),
+            "vehiculo_modelo":forms.NumberInput(attrs={
+                "class":"form-control",
                 }),
         }