+++ polizador/secretariador/models.py
@@ -281,6 +281,11 @@ class Vehiculo(models.Model):
         output_field=models.TextField(),
         db_persist=True,
     )
+    vehiculo_titular_agente = models.ForeignKey("Comisionado", on_delete=models.CASCADE, null=True, blank=True)
+    vehiculo_titular_empresa = models.ForeignKey("carga.Empresa", on_delete=models.CASCADE, null=True, blank=True)
+    vehiculo_n_motor = models.CharField("Número de Motor", max_length=100, null=True, blank=True)
+    vehiculo_n_chasis = models.CharField("Número de Chasis", max_length=100, null=True, blank=True)
+    vechiculo_modelo = models.DecimalField("Modelo", max_digits=4, decimal_places=0, null=True, blank=True)
 
     def __str__(self):
         return f"{self.vehiculo_modelo} - {self.vehiculo_patente}"